refactor: extract overlay instances logic to new mixin
